refactor(backend): extract markdown scanner from markdownMedia
Move the hand-written CommonMark scanner out of workspacePackages/markdownMedia.ts
into shared/markdown/ (complexity, lexer, blockScanner, inlineScanner). markdownMedia.ts
keeps its full export surface and now holds only the managed-media path API.
