+++ apps/backend/src/shared/markdown/index.ts
@@ -1,0 +1,5 @@
+export { MarkdownComplexityError } from "./complexity";
+export {
+  extractMarkdownNonCodeTextSegmentsUnchecked,
+  iterateMarkdownActiveDestinations,
+} from "./inlineScanner";
